refactor: split quic layer and h3 layer
