+++ rpxy-lib/src/proxy/mod.rs
@@ -3,6 +3,8 @@ mod proxy_client_cert;
 #[cfg(feature = "http3")]
 mod proxy_h3;
 mod proxy_main;
+#[cfg(feature = "http3")]
+mod proxy_quic;
 mod proxy_tls;
 mod socket;
